[refactor:move] apps/desktop: DesktopShellPage.test.tsx をテーマ別15ファイルへ分割
4,643行・112宣言(実行117)の単一テストファイルを、テーマ別15ファイル
(routing/notifications/shellChrome/topics/composer/reactions/
timelineRefresh/channels/mediaComposer/mediaRendering/socialGraph/
detailPanes/messages/workspaceResilience/profile)へ contiguous 行範囲の
verbatim 移動で分割。テスト本文・アサーションは無変更。

分割によりワーカー数が増えて高コアマシンで全画面レンダーテストが
CPU オーバーサブスクリプション起因の flake を起こすため、
vite.config.ts に test.maxWorkers: 6 を追加(分割前の実行モデルへの
上限キャップ。低コアCIでは実質 no-op)。

検証: 宣言112/実行117 一致、フルスイート305テスト3連続 green、
desktop-ui-check green、oversized-files 解消(全ファイル1000行未満)。

Refs: .claude/plans/2026-07-03-refactor-s1-test-split.md (T1 / 付録A)
