+++ apps/desktop/src/shell/DesktopShellPage.mediaRendering.test.tsx
@@ -1,0 +1,463 @@
+import { fireEvent, render, screen, waitFor, within } from '@testing-library/react';
+import userEvent from '@testing-library/user-event';
+import { afterEach, beforeEach, expect, test, vi } from 'vitest';
+
+import { createDesktopMockApi } from '@/mocks/desktopApiMock';
+import { App } from '@/App';
+import {
+  buildImagePost,
+  buildVideoPost,
+  installObjectUrlMocks,
+  setViewportWidth,
+} from './DesktopShellPage.testHelpers';
+import type { BlobViewStatus } from '@/lib/api';
+
+beforeEach(() => {
+  setViewportWidth(1024);
+  window.history.replaceState(null, '', '/');
+});
+
+afterEach(() => {
+  vi.useRealTimers();
+});
+
+test('timeline image post shows media skeleton when attachment is missing', async () => {
+  const api = createDesktopMockApi({
+    seedPosts: {
+      'kukuri:topic:demo': [buildImagePost()],
+    },
+  });
+  api.getBlobMediaPayload = async () => null;
+
+  render(<App api={api} />);
+
+  await waitFor(() => {
+    expect(screen.getByTestId('media-skeleton-image-post')).toBeInTheDocument();
+  });
+  expect(screen.getByTestId('media-skeleton-image-post')).toBeInTheDocument();
+  expect(screen.getByText('image/png')).toBeInTheDocument();
+});
+
+test('timeline image post switches to ready state when attachment becomes available', async () => {
+  const missingPost = buildImagePost();
+  const { rerender } = render(
+    <App
+      api={createDesktopMockApi({
+        seedPosts: {
+          'kukuri:topic:demo': [missingPost],
+        },
+      })}
+    />
+  );
+
+  await waitFor(() => {
+    expect(screen.getByTestId('media-skeleton-image-post')).toBeInTheDocument();
+  });
+
+  rerender(
+    <App
+      api={createDesktopMockApi({
+        seedPosts: {
+          'kukuri:topic:demo': [
+            buildImagePost({
+              content: 'caption ready',
+              content_status: 'Available' satisfies BlobViewStatus,
+              attachments: [
+                {
+                  ...missingPost.attachments[0],
+                  status: 'Available',
+                },
+              ],
+            }),
+          ],
+        },
+      })}
+    />
+  );
+
+  await waitFor(() => {
+    expect(screen.getByText('caption ready')).toBeInTheDocument();
+  });
+  expect(screen.queryByTestId('media-skeleton-image-post')).not.toBeInTheDocument();
+});
+
+test('timeline image post renders actual preview when object-url payload is available', async () => {
+  installObjectUrlMocks();
+  const api = createDesktopMockApi({
+    seedPosts: {
+      'kukuri:topic:demo': [
+        buildImagePost({
+          content: 'caption ready',
+          content_status: 'Available',
+          attachments: [
+            {
+              hash: 'b'.repeat(64),
+              mime: 'image/png',
+              bytes: 4096,
+              role: 'image_original',
+              status: 'Available',
+            },
+          ],
+        }),
+      ],
+    },
+  });
+  api.getBlobMediaPayload = async () => ({
+    bytes_base64: 'ZmFrZS1pbWFnZQ==',
+    mime: 'image/png',
+  });
+
+  render(<App api={api} />);
+
+  const preview = await screen.findByTestId('media-preview-image-post');
+  expect(preview).toBeInTheDocument();
+  expect(preview.getAttribute('src')).toContain('blob:mock-');
+});
+
+test('thread pane reuses the same image placeholder renderer', async () => {
+  const user = userEvent.setup();
+  const api = createDesktopMockApi({
+    seedPosts: {
+      'kukuri:topic:demo': [
+        buildImagePost(),
+        {
+          ...buildImagePost({
+            object_id: 'reply-post',
+            envelope_id: 'envelope-reply-post',
+            object_kind: 'comment',
+            content: 'reply body',
+            content_status: 'Available',
+            attachments: [],
+            reply_to: 'image-post',
+            root_id: 'image-post',
+          }),
+        },
+      ],
+    },
+  });
+  api.getBlobMediaPayload = async () => null;
+
+  render(
+    <App api={api} />
+  );
+
+  await waitFor(() => {
+    expect(screen.getByText('envelope-image-post')).toBeInTheDocument();
+  });
+  await user.click(screen.getByText('envelope-image-post'));
+  const threadPanel = await screen.findByRole('complementary', { name: 'Thread' });
+
+  await waitFor(() => {
+    expect(within(threadPanel).getByTestId('media-skeleton-image-post')).toBeInTheDocument();
+  });
+  expect(within(threadPanel).getByTestId('media-skeleton-image-post')).toBeInTheDocument();
+});
+
+test('text body pending uses text skeleton without hiding image metadata', async () => {
+  render(
+    <App
+      api={createDesktopMockApi({
+        seedPosts: {
+          'kukuri:topic:demo': [buildImagePost()],
+        },
+      })}
+    />
+  );
+
+  await waitFor(() => {
+    expect(screen.getByTestId('text-skeleton-image-post')).toBeInTheDocument();
+  });
+  expect(screen.getByText('image/png')).toBeInTheDocument();
+  expect(screen.getByText('2.0 KB')).toBeInTheDocument();
+});
+
+test('timeline video post shows poster skeleton when poster is missing', async () => {
+  const api = createDesktopMockApi({
+    seedPosts: {
+      'kukuri:topic:demo': [buildVideoPost()],
+    },
+  });
+  api.getBlobMediaPayload = async () => null;
+
+  render(<App api={api} />);
+
+  await waitFor(() => {
+    expect(screen.getByTestId('media-skeleton-video-post')).toBeInTheDocument();
+  });
+  expect(screen.getByTestId('media-skeleton-video-post')).toBeInTheDocument();
+  expect(screen.getByText('video/mp4')).toBeInTheDocument();
+});
+
+test('poster-only video card renders poster preview without video element', async () => {
+  installObjectUrlMocks();
+  const api = createDesktopMockApi({
+    seedPosts: {
+      'kukuri:topic:demo': [
+        buildVideoPost({
+          attachments: [
+            {
+              hash: 'v'.repeat(64),
+              mime: 'video/mp4',
+              bytes: 8192,
+              role: 'video_manifest',
+              status: 'Missing',
+            },
+            {
+              hash: 'p'.repeat(64),
+              mime: 'image/jpeg',
+              bytes: 1024,
+              role: 'video_poster',
+              status: 'Available',
+            },
+          ],
+        }),
+      ],
+    },
+  });
+  api.getBlobMediaPayload = async (hash, mime) =>
+    hash === 'p'.repeat(64)
+      ? {
+          bytes_base64: 'ZmFrZS1wb3N0ZXI=',
+          mime,
+        }
+      : null;
+
+  render(<App api={api} />);
+
+  const posterPreview = await screen.findByTestId('media-preview-video-post');
+  expect(posterPreview).toBeInTheDocument();
+  expect(screen.queryByTestId('media-video-video-post')).not.toBeInTheDocument();
+  expect(posterPreview.getAttribute('src')).toContain('blob:mock-');
+});
+
+test('video card fetches manifest payload even when attachment status is missing', async () => {
+  installObjectUrlMocks();
+  const api = createDesktopMockApi({
+    seedPosts: {
+      'kukuri:topic:demo': [
+        buildVideoPost({
+          attachments: [
+            {
+              hash: 'late-manifest'.repeat(4),
+              mime: 'video/mp4',
+              bytes: 9999,
+              role: 'video_manifest',
+              status: 'Missing',
+            },
+            {
+              hash: 'late-poster'.repeat(4),
+              mime: 'image/jpeg',
+              bytes: 1024,
+              role: 'video_poster',
+              status: 'Available',
+            },
+          ],
+        }),
+      ],
+    },
+  });
+  api.getBlobMediaPayload = async (hash, mime) => {
+    if (hash === 'late-manifest'.repeat(4)) {
+      return {
+        bytes_base64: 'ZmFrZS12aWRlbw==',
+        mime,
+      };
+    }
+    if (hash === 'late-poster'.repeat(4)) {
+      return {
+        bytes_base64: 'ZmFrZS1wb3N0ZXI=',
+        mime,
+      };
+    }
+    return null;
+  };
+
+  render(<App api={api} />);
+
+  const video = await screen.findByTestId('media-video-video-post');
+  expect(video).toBeInTheDocument();
+  expect(video).toHaveAttribute('src', expect.stringContaining('blob:mock-'));
+});
+
+test('video card retries after stalled manifest fetch after rerender', async () => {
+  installObjectUrlMocks();
+  const manifestHash = 'retry-manifest'.repeat(4);
+  const posterHash = 'retry-poster'.repeat(4);
+  const seedPosts = {
+    'kukuri:topic:demo': [
+      buildVideoPost({
+        attachments: [
+          {
+            hash: manifestHash,
+            mime: 'video/mp4',
+            bytes: 9999,
+            role: 'video_manifest',
+            status: 'Missing',
+          },
+          {
+            hash: posterHash,
+            mime: 'image/jpeg',
+            bytes: 1024,
+            role: 'video_poster',
+            status: 'Missing',
+          },
+        ],
+      }),
+    ],
+  };
+  const stalledApi = createDesktopMockApi({
+    seedPosts,
+  });
+  stalledApi.getBlobMediaPayload = async (hash, mime) => {
+    if (hash === manifestHash) {
+      return new Promise<null>(() => {});
+    }
+    if (hash === posterHash) {
+      return {
+        bytes_base64: 'ZmFrZS1wb3N0ZXI=',
+        mime,
+      };
+    }
+    return null;
+  };
+  const recoveredApi = createDesktopMockApi({
+    seedPosts: {
+      ...seedPosts,
+    },
+  });
+  recoveredApi.getBlobMediaPayload = async (hash, mime) => {
+    if (hash === manifestHash) {
+      return {
+        bytes_base64: 'ZmFrZS12aWRlbw==',
+        mime,
+      };
+    }
+    if (hash === posterHash) {
+      return {
+        bytes_base64: 'ZmFrZS1wb3N0ZXI=',
+        mime,
+      };
+    }
+    return null;
+  };
+
+  const { rerender } = render(<App api={stalledApi} />);
+
+  await waitFor(() => {
+    expect(screen.getByTestId('media-preview-video-post')).toBeInTheDocument();
+  });
+
+  rerender(<App api={recoveredApi} />);
+
+  const video = await screen.findByTestId('media-video-video-post');
+  expect(video).toBeInTheDocument();
+  expect(video).toHaveAttribute('src', expect.stringContaining('blob:mock-'));
+});
+
+test('video card renders object-url playback source when manifest payload is available', async () => {
+  installObjectUrlMocks();
+  const api = createDesktopMockApi({
+    seedPosts: {
+      'kukuri:topic:demo': [
+        buildVideoPost({
+          attachments: [
+            {
+              hash: 'manifest'.repeat(8),
+              mime: 'video/mp4',
+              bytes: 9999,
+              role: 'video_manifest',
+              status: 'Available',
+            },
+            {
+              hash: 'poster'.repeat(8),
+              mime: 'image/jpeg',
+              bytes: 1024,
+              role: 'video_poster',
+              status: 'Available',
+            },
+          ],
+        }),
+      ],
+    },
+  });
+  api.getBlobMediaPayload = async (hash, mime) => {
+    if (hash === 'manifest'.repeat(8)) {
+      return {
+        bytes_base64: 'ZmFrZS12aWRlbw==',
+        mime,
+      };
+    }
+    if (hash === 'poster'.repeat(8)) {
+      return {
+        bytes_base64: 'ZmFrZS1wb3N0ZXI=',
+        mime,
+      };
+    }
+    return null;
+  };
+
+  render(<App api={api} />);
+
+  const video = await screen.findByTestId('media-video-video-post');
+  expect(video).toBeInTheDocument();
+  expect(video.getAttribute('src')).toContain('blob:mock-');
+});
+
+test('video card falls back to poster preview when playback is unsupported on this client', async () => {
+  installObjectUrlMocks();
+  const api = createDesktopMockApi({
+    seedPosts: {
+      'kukuri:topic:demo': [
+        buildVideoPost({
+          attachments: [
+            {
+              hash: 'manifest'.repeat(8),
+              mime: 'video/mp4',
+              bytes: 9999,
+              role: 'video_manifest',
+              status: 'Available',
+            },
+            {
+              hash: 'poster'.repeat(8),
+              mime: 'image/jpeg',
+              bytes: 1024,
+              role: 'video_poster',
+              status: 'Available',
+            },
+          ],
+        }),
+      ],
+    },
+  });
+  api.getBlobMediaPayload = async (hash, mime) => {
+    if (hash === 'manifest'.repeat(8)) {
+      return {
+        bytes_base64: 'ZmFrZS12aWRlbw==',
+        mime,
+      };
+    }
+    if (hash === 'poster'.repeat(8)) {
+      return {
+        bytes_base64: 'ZmFrZS1wb3N0ZXI=',
+        mime,
+      };
+    }
+    return null;
+  };
+
+  render(<App api={api} />);
+
+  const video = await screen.findByTestId('media-video-video-post');
+  Object.defineProperty(video, 'error', {
+    configurable: true,
+    get: () => ({ code: 4 }),
+  });
+  fireEvent.error(video);
+
+  await waitFor(() => {
+    expect(screen.queryByTestId('media-video-video-post')).not.toBeInTheDocument();
+  });
+  expect(screen.getByTestId('media-preview-video-post')).toBeInTheDocument();
+  expect(screen.getByAltText('video poster')).toBeInTheDocument();
+});
+
